Optimized Element.resolvePrefix by moving one lookup out of inner loop.
It really make no sense to look into the class variable specialprefixes
on each iteration. As the xmlns prefix should not be redefined, we can
also look there first.

For my load case, 10 % of the time of analyzing the reply from the server
is spend in that method. I would really like to cache the namespace
prefixes, but maintaining the caches turned out to be too expensive.

(taken from https://bitbucket.org/bluehorn/suds-performance/)