Optimize creation of Text instances.
For empty instances w/o lang and escaped assigned a single instance is created and
returned instead of creating new instances.

For the common case of not passing lang and escaped flags, just the unicode
constructor is called. This saves assignment to the lang and escaped fields,
which are modelled as properties now.

(taken from https://bitbucket.org/bluehorn/suds-performance/)